% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcode2html.R
\name{Rcode2html}
\alias{Rcode2html}
\title{Highlight R Code with Custom Styles}
\usage{
Rcode2html(
  code,
  style = 1,
  color_backgroud = "#111111",
  color_function = "#B294BB",
  color_operator = "#8ABEB7",
  color_number = "#DE935F",
  color_quotation = "#F0C674",
  color_comments = "#969896",
  color_others = "#FFFFFF"
)
}
\arguments{
\item{code}{A character string representing the R code to be highlighted.
Inside the code, the inner quotes should be escaped with backslashes to prevent them from prematurely ending the string in R (see example).}

\item{style}{An integer indicating the highlighting style, with options 1, 2, or 3. Default is `1` (recommend).
Different styles may apply variations in background frame, font, underline, or other
decorations (future expansion).}

\item{color_backgroud}{A string specifying the background color of the code block.
Default is `"#111111"` (dark background).}

\item{color_function}{A string specifying the color for function names.
Default is `"#B294BB"` (purple shade).}

\item{color_operator}{A string specifying the color for operators (`+`, `-`, etc.).
Default is `"#8ABEB7"` (cyan shade).}

\item{color_number}{A string specifying the color for numeric literals.
Default is `"#DE935F"` (orange shade).}

\item{color_quotation}{A string specifying the color for quoted strings.
Default is `"#F0C674"` (yellow shade).}

\item{color_comments}{A string specifying the color for comments (lines starting with `#`).
Default is `"#969896"` (gray shade).}

\item{color_others}{A string specifying the default color for all other elements.
Default is `"#FFFFFF"` (white).}
}
\value{
A character string with HTML-style span elements for syntax highlighting.
        The returned string can be directly rendered in an HTML environment.
}
\description{
This function offers syntax highlighting for R code, transforming plain script into visually appealing HTML-formatted code blocks with customizable color palettes.
It supports highlighting different elements of R syntax, such as functions, operators, numbers, comments, and others.
The returned string can be directly rendered in an HTML environment.
}
\examples{
code <- "x <- 11 # Assign value to x
y <- \"eleven\" # Assign character to y"
cat(Rcode2html(code))
cat(Rcode2html(code, color_backgroud = "#323232",color_operator = "#00BFFF",
               color_number="#A52A2A", color_quotation = "#CDAD00", color_comments="#6495ED"))


code <- "# use package car
library(car)
m <- colMeans(x)
S <- cov(x)
plot(x[,1], x[,2], pch=19, xlim=c(-2.5,9.5), ylim=c(-2.5,9.5))
car::ellipse(m, S, 3.03, col=\"brown\", center.cex=1, lwd=3)"
cat(Rcode2html(code))
cat(Rcode2html(code, color_function = "#FF0000"))

}
